fix: lazy imports, linter settings
