+++ pyproject.toml
@@ -20,6 +20,9 @@ members = [
     "images/uvbase"
 ]
 
+[tool.ruff]
+line-length = 120
+
 [tool.uv]
 dev-dependencies = [
     "jupyter_core>=5.0.0",
@@ -34,6 +37,32 @@ dev-dependencies = [
     "parameterized>=0.9.0",
     "tomli-w>=1.2.0",
 ]
+
+[tool.ruff.lint]
+preview = true
+select = [
+    "E",  # pycodestyle
+    "F",  # pyflakes
+    "UP",  # pyupgrade
+    "B",  # flake8-bugbear
+    "SIM",  # flake8-simplify
+    "I",  # isort
+    "PLC0415",  # import-outside-toplevel - disallow lazy imports
+]
+ignore = [
+    # Ignore other preview rules we're not ready for yet
+    "UP042",  # Replace str/Enum with StrEnum (Python 3.11+)
+    "UP046",  # Use PEP 695 type parameters (Python 3.12+)
+    "UP047",  # Use PEP 695 type parameters for functions (Python 3.12+)
+]
+
+[tool.ruff.lint.per-file-ignores]
+# Allow lazy imports in provider factory (handles optional dependencies)
+"**/provider/instruction_runner_factory.py" = ["PLC0415"]
+# Allow lazy imports in tests (common pattern for imports after mocks)
+"tests/**/*.py" = ["PLC0415"]
+# GitHub utility scripts - allow simpler code patterns
+".github/**/*.py" = ["SIM108"]
 
 [tool.mypy]
 warn_return_any = true
